backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
ollama
pydantic
langfuse


# LangChain dependencies
langchain
langchain-community
langchain-core



# For RAG
chromadb # Keep for local backup if needed
tiktoken
pypdf
langchain-ollama
pandas
openpyxl
sqlalchemy
python-docx

# Optional but recommended
redis  # For production memory management
langgraph
requests
beautifulsoup4